+++ cli/src/cli_util.rs
@@ -71,6 +71,8 @@ use jj_lib::fileset::FilesetExpression;
 use jj_lib::gitignore::GitIgnoreError;
 use jj_lib::gitignore::GitIgnoreFile;
 use jj_lib::id_prefix::IdPrefixContext;
+use jj_lib::mailmap::read_current_mailmap;
+use jj_lib::mailmap::Mailmap;
 use jj_lib::matchers::Matcher;
 use jj_lib::merge::MergedTreeValue;
 use jj_lib::merged_tree::MergedTree;
@@ -137,6 +139,7 @@ use jj_lib::workspace::Workspace;
 use jj_lib::workspace::WorkspaceLoadError;
 use jj_lib::workspace::WorkspaceLoader;
 use jj_lib::workspace::WorkspaceLoaderFactory;
+use pollster::FutureExt;
 use tracing::instrument;
 use tracing_chrome::ChromeLayerBuilder;
 use tracing_subscriber::prelude::*;
@@ -843,6 +846,8 @@ impl WorkspaceCommandEnvironment {
             date_pattern_context: now.into(),
             extensions: self.command.revset_extensions(),
             workspace: Some(workspace_context),
+            // TODO: Consider handling errors here.
+            mailmap: Rc::new(self.current_mailmap().unwrap_or_default()),
         }
     }
 
@@ -1327,6 +1332,14 @@ to the current parents may contain changes from multiple commits.
 
     pub fn get_wc_commit_id(&self) -> Option<&CommitId> {
         self.repo().view().get_wc_commit_id(self.workspace_name())
+    }
+
+    pub fn current_mailmap(&self) -> Result<Mailmap, CommandError> {
+        // TODO: Consider figuring out a caching strategy for this.
+        Ok(
+            read_current_mailmap(self.repo().as_ref(), self.workspace.workspace_name())
+                .block_on()?,
+        )
     }
 
     pub fn working_copy_shared_with_git(&self) -> bool {