+++ lib/src/lib.rs
@@ -77,6 +77,7 @@ pub mod id_prefix;
 pub mod index;
 pub mod local_working_copy;
 pub mod lock;
+pub mod mailmap;
 pub mod matchers;
 pub mod merge;
 pub mod merged_tree;